refactor: PostsManagerPage 경로 변경
